AF1121-1
当たり判定をまとめるのではなく各自で持てるようにした

+++ ArcherFish/Src/Application/Object/TextDraw/TextDraw.cpp
@@ -2,7 +2,7 @@
 
 void TextDraw::DrawSprite()
 {
-	KdShaderManager::Instance().m_spriteShader.DrawTex(&m_tex, m_pixel.x, m_pixel.y, &m_recAngle, &m_color);
+	KdShaderManager::Instance().m_spriteShader.DrawTex(&m_tex, (int)m_pixel.x, (int)m_pixel.y, &m_recAngle, &m_color);
 }
 
 void TextDraw::SetLoadText(const std::string _tex, const Math::Vector2 _pixel, const Math::Rectangle _recAng)
